Revert "OS-3977: Support min_platform requirement in imgadm"
This reverts commit 5da0f52723652140dc3446e47d12fecb30a5b1c6.

Reverts bug introduced in latest SmartOS that prevents the import of
newer images.

On SmartOS imgadm mistakes
it for an old version of SDC and as such won't allow images to be used
that require a newer platform. This check was never present in smartos
before only SDC. So for now we just revert it and carry on as we were :)